pip-requirements.txt
psutil
bottle
beaker
bottle-cork
blinker
requests
lxml
webtest
pyyaml
prov
flask
flask-sqlalchemy
flask-security
flask-admin
bcrypt
graphviz
pydot
pydotplus
pytest